refactor analytics to use tracks package
