+++ client/lib/analytics/utils/index.js
@@ -3,7 +3,6 @@
  */
 export { default as MARKETING_COUPONS_KEY } from './marketing-coupons-key';
 export { default as costToUSD } from './cost-to-usd';
-export { default as doNotTrack } from './do-not-track';
 export { default as hashPii } from './hash-pii';
 export { default as isPiiUrl } from './is-pii-url';
 export { default as isUrlBlacklistedForPerformance } from './is-url-blacklisted-for-performance';
