+++ client/lib/analytics/utils/is-ad-tracking-allowed.js
@@ -3,11 +3,10 @@
  */
 import config from 'config';
 import debug from './debug';
-import doNotTrack from './do-not-track';
 import isUrlBlacklistedForPerformance from './is-url-blacklisted-for-performance';
 import isPiiUrl from './is-pii-url';
 import mayWeTrackCurrentUserGdpr from './may-we-track-current-user-gdpr';
-
+import { doNotTrack } from '@automattic/calypso-analytics';
 /**
  * Returns whether ad tracking is allowed.
  *
