+++ packages/calypso-analytics/src/index.js
@@ -1,7 +1,10 @@
 /**
  * External dependencies
  */
-import { includes, isObjectLike, isUndefined, omit } from 'lodash';
+import { assign, includes, isObjectLike, isUndefined, omit, pickBy, times } from 'lodash';
+import { parse } from 'qs';
+import cookie from 'cookie';
+import url from 'url';
 import debug from 'debug';
 import { EventEmitter } from 'events';
 
@@ -9,7 +12,14 @@ import { EventEmitter } from 'events';
  * Internal Dependencies
  */
 
-import { setCurrentUser } from './utils/current-user';
+import { getCurrentUser, setCurrentUser } from './utils/current-user';
+import { loadScript } from '@automattic/load-script';
+import doNotTrack from './utils/do-not-track';
+
+/**
+ * Re-export
+ */
+export { default as doNotTrack } from './utils/do-not-track';
 
 /**
  * Module variables
@@ -17,6 +27,7 @@ import { setCurrentUser } from './utils/current-user';
 const initializeDebug = debug( 'calypso:analytics:initialize' );
 const tracksDebug = debug( 'calypso:analytics:tracks' );
 const identifyUserDebug = debug( 'calypso:analytics:identifyUser' );
+const blockedTracksDebug = debug( 'calypso:analytics:blockedTracks' );
 
 /**
  * Tracks uses a bunch of special query params that should not be used as property name
@@ -33,13 +44,84 @@ const EVENT_NAME_EXCEPTIONS = [
 	'wcadmin_storeprofiler_payment_create_account',
 ];
 let _superProps; // Added to all Tracks events.
+let _loadTracksResult = Promise.resolve(); // default value for non-BOM environments.
 
-// Load tracking scripts
 if ( typeof window !== 'undefined' ) {
 	window._tkq = window._tkq || [];
 }
 
+if ( typeof document !== 'undefined' ) {
+	_loadTracksResult = loadScript( '//stats.wp.com/w.js?60' );
+}
+
+function getUrlParameter( name ) {
+	name = name.replace( /[[]/g, '\\[' ).replace( /[\]]/g, '\\]' );
+	const regex = new RegExp( '[\\?&]' + name + '=([^&#]*)' );
+	const results = regex.exec( window.location.search );
+	return results === null ? '' : decodeURIComponent( results[ 1 ].replace( /\+/g, ' ' ) );
+}
+
+function createRandomId( randomBytesLength = 9 ) {
+	// 9 * 4/3 = 12
+	// this is to avoid getting padding of a random byte string when it is base64 encoded
+	let randomBytes;
+
+	if ( window.crypto && window.crypto.getRandomValues ) {
+		randomBytes = new Uint8Array( randomBytesLength );
+		window.crypto.getRandomValues( randomBytes );
+	} else {
+		randomBytes = times( randomBytesLength, () => Math.floor( Math.random() * 256 ) );
+	}
+
+	return window.btoa( String.fromCharCode.apply( String, randomBytes ) );
+}
+
+function checkForBlockedTracks() {
+	// Proceed only after the tracks script load finished and failed.
+	// Calling this function from `initialize` ensures current user is set.
+	// This detects stats blocking, and identifies by `getCurrentUser()`, URL, or cookie.
+	_loadTracksResult.catch( () => {
+		let _ut, _ui;
+		const currentUser = getCurrentUser();
+
+		if ( currentUser && currentUser.ID ) {
+			_ut = 'wpcom:user_id';
+			_ui = currentUser.ID;
+		} else {
+			_ut = getUrlParameter( '_ut' ) || 'anon';
+			_ui = getUrlParameter( '_ui' );
+
+			if ( ! _ui ) {
+				const cookies = cookie.parse( document.cookie );
+				if ( cookies.tk_ai ) {
+					_ui = cookies.tk_ai;
+				} else {
+					const randomIdLength = 18; // 18 * 4/3 = 24 (base64 encoded chars).
+					_ui = createRandomId( randomIdLength );
+					document.cookie = cookie.serialize( 'tk_ai', _ui );
+				}
+			}
+		}
+
+		blockedTracksDebug( 'Loading /nostats.js', { _ut, _ui } );
+		return loadScript(
+			'/nostats.js?_ut=' + encodeURIComponent( _ut ) + '&_ui=' + encodeURIComponent( _ui )
+		);
+	} );
+}
+
 export const analyticsEvents = new EventEmitter();
+
+/**
+ * Returns the anoymous id stored in the `tk_ai` cookie
+ *
+ * @returns {string} - The Tracks anonymous user id
+ */
+export function getTracksAnonymousUserId() {
+	const cookies = cookie.parse( document.cookie );
+
+	return cookies.tk_ai;
+}
 
 export function initializeAnalytics( currentUser, superProps ) {
 	// Update super props.
@@ -51,7 +133,7 @@ export function initializeAnalytics( currentUser, superProps ) {
 	// Identify current user.
 	if ( 'object' === typeof currentUser ) {
 		initializeDebug( 'identifyUser', currentUser );
-		analytics.identifyUser( currentUser );
+		identifyUser( currentUser );
 	}
 
 	// Tracks blocked?
@@ -78,7 +160,7 @@ export function identifyUser( userData ) {
 	window._tkq.push( [ 'identifyUser', currentUser.ID, currentUser.username ] );
 }
 
-export function recordEvent( eventName, eventProperties ) {
+export function recordTracksEvent( eventName, eventProperties ) {
 	eventProperties = eventProperties || {};
 
 	if ( process.env.NODE_ENV !== 'production' && typeof console !== 'undefined' ) {
@@ -149,4 +231,29 @@ export function recordEvent( eventName, eventProperties ) {
 		}
 		analyticsEvents.emit( 'record-event', eventName, eventProperties );
 	}
+}
+
+export function recordTracksPageView( urlPath, params ) {
+	let eventProperties = {
+		build_timestamp: window.BUILD_TIMESTAMP,
+		do_not_track: doNotTrack() ? 1 : 0,
+		path: urlPath,
+	};
+
+	// add optional path params
+	if ( params ) {
+		eventProperties = assign( eventProperties, params );
+	}
+
+	// Record all `utm` marketing parameters as event properties on the page view event
+	// so we can analyze their performance with our analytics tools
+	if ( window.location ) {
+		const parsedUrl = url.parse( window.location.href );
+		const urlParams = parse( parsedUrl.query );
+		const utmParams = pickBy( urlParams, ( value, key ) => key.startsWith( 'utm_' ) );
+
+		eventProperties = assign( eventProperties, utmParams );
+	}
+
+	recordTracksEvent( 'calypso_page_view', eventProperties );
 }
