+++ packages/calypso-analytics/src/utils/debug.js
@@ -1,0 +1,9 @@
+/**
+ * External dependencies
+ */
+import debugFactory from 'debug';
+
+/**
+ * Module variables
+ */
+export default debugFactory( 'calypso:analytics:utils' );
